+++ src/system/Zikula/Module/SearchModule/SearchModuleInstaller.php
@@ -40,9 +40,15 @@ class SearchModuleInstaller extends \Zikula_AbstractInstaller
         // create module vars
         $this->setVar('itemsperpage', 10);
         $this->setVar('limitsummary', 255);
+        $this->setVar('opensearch_enabled', true);
+        $this->setVar('opensearch_adult_content', false);
 
-        // register event handler to activate new modules in the search block
-        EventUtil::registerPersistentModuleHandler('ZikulaSearchModule', 'installer.module.installed', array('Zikula\Module\SearchModule\Listener\ModuleListener', 'moduleInstall'));
+
+        // register event handler to activate new modules in the search block.
+        EventUtil::registerPersistentModuleHandler($this->name, 'installer.module.installed', array('Zikula\Module\SearchModule\Listener\ModuleListener', 'moduleInstall'));
+
+        // register event handler for opensearch.
+        EventUtil::registerPersistentModuleHandler($this->name, 'frontcontroller.predispatch', array('Zikula\Module\SearchModule\Listener\PageloadListener', 'pageload'));
 
         // Initialisation successful
         return true;
@@ -53,6 +59,10 @@ class SearchModuleInstaller extends \Zikula_AbstractInstaller
         // Upgrade dependent on old version number
         switch ($oldversion) {
             case '1.5.2':
+                $this->setVar('opensearch_enabled', true);
+                $this->setVar('opensearch_adult_content', false);
+                EventUtil::registerPersistentModuleHandler($this->name, 'frontcontroller.predispatch', array('Zikula\Module\SearchModule\Listener\PageloadListener', 'pageload'));
+            case '1.5.3':
             // future upgrade routines
         }