+++ src/docs/CHANGELOG.md
@@ -63,6 +63,7 @@ Features:
 - Reset start page module to static frontpage if it is deactivated, #104.
 - jQuery and jQuery UI are now outsourced to their own bundles.
 - Added events if a module is activated and if a module is deactivated.
+- Implemented OpenSearch.
 
 CHANGELOG - ZIKULA 1.3.5
 ------------------------